+++ src/middleware/gaspriceMiddleware.js
@@ -1,7 +1,7 @@
 const Web3 = require('web3')
 const maxGaspriceGWei = 500
 
-const gaspriceMiddleware = ({verbose, gasPrice}) => {
+const gaspriceMiddleware = ({verbose, gasPrice, requiresAccount}) => {
     // check limits
     if (gasPrice > maxGaspriceGWei) {
         throw Error(`Gas price too high. Maximum possible value is ${maxGaspriceGWei} gwei (provided value: ${gasPrice})`)
@@ -9,7 +9,11 @@ const gaspriceMiddleware = ({verbose, gasPrice}) => {
     // convert to wei BN instance
     const gwei = Web3.utils.toBN(gasPrice)
     const wei = Web3.utils.toWei(gwei, 'gwei')
-    verbose && console.log(`Setting gas price: ${gasPrice} gwei`)
+
+    // ignore gas price when command does not require account and no tx will be performed
+    if (requiresAccount) {
+        verbose && console.log(`Setting gas price: ${gasPrice} gwei`)
+    }
     return {
         gasPrice: wei
     }